% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_ratios.R
\name{all_ratios}
\alias{all_ratios}
\title{Complete processing for Pb isotope ratio data.}
\usage{
all_ratios(filename, report = c("short", "long"), print = c(FALSE, TRUE))
}
\arguments{
\item{filename}{A dataset saved in csv format with "," as field separator and "." as
decimal separator. Such dataset is derived from .rep files provided by the
instrument by opening it with spreadsheet editor and saving it as .csv file using
a local setting = EN-US or equivalent.
The .csv file must contain the sample name, a summary of the measuring conditions,
the raw intensities for each monitored isotopes and a final summary with average
signal intensities and isotope ratios. These pieces of information are repeated
for each sample. Bracketing standards are assumed to be NIST SRM 981 and should
be labelled as \code{SRM}. QC samples are assumed to be BCR CRM 482 and should be
labbelled as \code{CRM}.}

\item{report}{A string referring to the type of report generated by the instrument
accepted values are either \code{"short"} or \code{"long"}.
\itemize{
\item{\bold{Short}}{ reports have only information about the sample ID,
sample time, a summary with average intensities and then the intensity values
for each replicate.}
\item{\bold{Long}}{ reports contain also information about the measurement conditions
 and the summary section is at the end of the intensity values for each replicates.
 They are usually produced when working in isotope ratio mode.}
}}

\item{print}{A logical string. \code{print = TRUE} will result in printing on screen
the ratios. \code{print = FALSE} is the default setting and it is suggested for
long datasets.}
}
\value{
A data.table containing Pb isotope ratios filtered for outliers and corrected
    for mass bias. \code{Pb20x20y} columns provide Pb ratio values for 20xPb/20yPb
    with \eqn{x, y = 6, 7, 8}. \code{Pb20x20y.U} columns stores the estimated extended
    uncertainty (\eqn{U = k u, k = 2}).
}
\description{
The function extracts, remove outliers and correct for mass bias Pb isotope ratio data
included in a .csv report generated by Perkin Elmer ELAN ICP-MS.
}
\examples{
file.short <- system.file("extdata", "sednya_2015.csv", package = "pbratios")
sednya.2015 <- all_ratios(file.short, report = "short", print = TRUE)

}
\seealso{
\code{\link{extract_data}}
  \code{\link{calc_ratios}}
  \code{\link{corr_mbf}}
}
