% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ratios.R
\name{calc.ratios}
\alias{calc.ratios}
\title{Removing outliers and averaging the replicates.}
\usage{
calc.ratios(data)
}
\arguments{
\item{data}{A data.table produced by the \code{extract.data} function, containing
raw Pb isotope data. The dataset must have a column "sample" containing the sample
name, the columns \code{Pb208207}, \code{Pb206207}, \code{Pb208206} and
\code{Pb207206} containing the measured Pb isotope ratio values for
208Pb/207Pb, 206Pb/207Pb, 208Pb/206Pb and 207Pb/206Pb, respectively.}
}
\value{
A data.table with mean Pb isotope ratio values for each sample. The number of
  replicates outside the \eqn{median ± 2.5 mad} range for each sample are printed on
  screen.
  Additionally, each average Pb isotope ratio is corredated with the experimental
  standard error which is stored in the columns \code{Pb20x20y.se}.
}
\description{
The function inspect raw Pb isotope ratios for outliers and summarise the replicates.
Data outside the \eqn{median ± 2.5 mad} criterion are automatically removed from the
dataset before computing the average ratios.
}
\examples{
file <- system.file("extdata", "spmnya_2012.csv", package = "pbratios")
spmnya.2012 <- extract.data(file)
spmnya.2012avg <- calc.ratios(spmnya.2012)

}
\seealso{
\code{\link{extract.data}}
}
