% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_massbias.R
\name{corr.mbf}
\alias{corr.mbf}
\title{Correcting for instrumental mass bias.}
\usage{
corr.mbf(data)
}
\arguments{
\item{data}{A data.table produced by the calc.ratios function, containing average Pb
isotope data. The dataset must have a column \code{sample} containing the sample
name, the columns \code{Pb208207}, \code{Pb206207}, \code{Pb208206} and
\code{Pb207206} containing the measured Pb isotope ratio values for 208Pb/207Pb,
206Pb/207Pb, 208Pb/206Pb and 207Pb/206Pb, respectively. For each ratio also a
column \code{Pb20x20y.se} containing the standard error is requested.
The data.table should start with a \code{SRM} sample, however, as it is considered
good practice to start the experimental sequence with the measurement of a blank
sample, the removal of such sample has to be performed manually before running
the \code{corr.mbf} function. This can be achieved
by using
\itemize{
\item{\code{data <- data[-grep("NO3$", sample), ]}}{ for blank samples labelled as
\code{"1\%HNO3"};}
\item{\code{data <- data[-grep("BK$", sample), ]}}{ for blank samples labelled as
\code{"BK"}.}
}}
}
\value{
a data.table containing the Pb isotope ratios corrected for the mass bias.
  The correction method is the bracketing technique. Bracketing standards should be
  labelled as \code{SRM} followed by a progressive integer (the calc.ratios function
  will rename the SRM samples accordingly). Each Pb isotope ratios is also corredated
  with the extended uncertainty. These values are stored in the columns
  \code{Pb20x20y.U} and are calculated by error propagation taking into account the
  sample standard error, the mass bias factor standard error and the uncertainty in
  the certified values for NIST SRM 981.
  \bold{Attention:} no warning must be produced by the function.
}
\description{
The function corrects Pb isotope ratio for mass bias using the standard bracketing
approach.
}
\examples{
file <- system.file("extdata", "spmnya_2012.csv", package = "pbratios")
spmnya.2012 <- extract.data(file)
spmnya.2012avg <- calc.ratios(spmnya.2012)
spmnya.2012cor <- corr.mbf(spmnya.2012avg)

}
\seealso{
\code{\link{extract.data}}
\code{\link{calc.ratios}}
}
