% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_massbias.R
\name{corr_mbf}
\alias{corr_mbf}
\title{Correcting instrumental mass bias.}
\usage{
corr_mbf(data)
}
\arguments{
\item{data}{A data.table produced by the calc.ratios function and containing average Pb isotope
data. The dataset must have a column \code{sample} containing the sample ID, the columns
\code{Pb208207}, \code{Pb206207}, \code{Pb208206} and \code{Pb207206} containing the measured
Pb isotope ratio values for 208Pb/207Pb, 206Pb/207Pb, 208Pb/206Pb and 207Pb/206Pb,
respectively. For each ratio also a columns \code{Pb20x20y.se}, containing standard errors are
required. Bracketing standards must be labelled as “SRM981” and are assumed to be diluted
solutions of NIST SRM 981. Certificate values for these ratios are reported at
\url{https://www-s.nist.gov/srmors/certificates/view_certPDF.cfm?certificate=981}. The function
will evaluate data from the first to the last \code{SRM} and samples outside these range will
be ignored.}
}
\value{
A data.table containing the Pb isotope ratios corrected for the mass bias using the
  standard bracketing technique. Extended uncertainty (\eqn{U = k u}, \eqn{k = 2}) is also
  provided for each ratio and it is stored in \code{Pb20x20y.U} columns. Uncertainty budget take
  into account the relative standard errors for the sample, the mass bias factor and the
  certified values for NIST SRM 981.
}
\description{
The function corrects Pb isotope ratio for mass bias using the standard bracketing approach.
}
\details{
The standard bracketing techniques uses the average mass bias factor measured in standard
solutions with known isotope ratio values to correct the isotope ratios measured in samples. As
mass bias factor is influenced by temporal random fluctuations, standard solutions must be
analysed just before and after the samples for which the correction will be applied.
}
\examples{
file.long <- system.file("extdata", "spmnya_2012.csv", package = "pbratios")
spmnya.2012 <- extract_data(file.long, report = "long")
spmnya.2012avg <- calc_ratios(spmnya.2012)
spmnya.2012cor <- corr_mbf(spmnya.2012avg)

}
\seealso{
\code{\link{extract_data}} \code{\link{calc_ratios}}
}
