% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{count.rows}
\alias{count.rows}
\title{Counting the rows of each report in the raw .csv file}
\usage{
count.rows(data = data)
}
\arguments{
\item{data}{a dataset saved in csv format with "," as field separator and "." as
decimal separator. Such dataset is derived from .rep files provided by the
instrument by opening it with spreadsheet editor and saving it as .csv file
using a local setting = EN-US or equivalent.
The .csv file must contain the sample name, a summary of the measuring conditions,
the raw intensities for each monitored isotopes and a final summary with average
signal intensities and isotope ratios. These pieces of information are repeated
for each sample.
The report start with "Isotope Ratio Report", whereas the final summary is
announced by "Summary".}
}
\value{
a list containing
   \itemize{
   \item{report.rows}{ number of rows occupied by each report;}
   \item{fullsummary}{ number of rows occupied by the summary.}
   }
}
\description{
The function counts the rows between two consecutive "Isotope Ratio Report" and the
rows required by the final "Summary" for each sample. This function is used internally
by extract.data.
}
