% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract.data}
\alias{extract.data}
\title{Extracting data from csv files produced by Perkin Elmer Elan ICP-MS.}
\usage{
extract.data(filename, report = c("short", "long"))
}
\arguments{
\item{filename}{A dataset saved in csv format with "," as field separator and "." as
decimal separator. Such dataset is derived from .rep files provided by the
instrument by opening it with spreadsheet editor and saving it as .csv file using
a local setting = EN-US or equivalent.
The .csv file must contain the sample name, a summary of the measuring conditions,
the raw intensities for each monitored isotopes and a final summary with average
signal intensities and isotope ratios. These pieces of information are repeated
for each sample. Bracketing standards are assumed to be NIST SRM 981 and should
be labelled as \code{SRM}. QC samples are assumed to be BCR CRM 482 and should be
labbelled as \code{CRM}.}

\item{report}{A string referring to the type of report generated by the instrument
accepted values are either \code{"short"} or \code{"long"}. \bold{Short} reports
have only information about the sample ID, sample time, a summary with average
intensities and then the intensity values for each replicate. \bold{Long} reports
contain also information about the measurement conditions and the summary section
is at the end of the intensity values for each replicates. \bold{Long} reports
are usually produced when working in isotope ratio mode.}
}
\value{
A data.table containing raw intensities and raw Pb isotope ratios
    for each sample. Provided isotope ratios are 208Pb/207Pb, 206Pb/207Pb, 208Pb/206Pb
    and 207Pb/206Pb.
    Note that these ratios have still to be inspected for outliers, summarised and
    corrected for mass bias.
}
\description{
Extracts 208Pb/207Pb, 206Pb/207Pb, 208Pb/206Pb and 207Pb/206Pb ratio data from the
raw .csv file produced by the instrument Perkin Elmer Elan ICP-MS working in isotope
ratio mode.
}
\examples{
file <- system.file("extdata", "spmnya_2012.csv", package = "pbratios")
spmnya.2012 <- extract.data(file)

}
