% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{Extracting raw Pb isotope ratio.}
\usage{
extract_data(filename, report = c("short", "long"))
}
\arguments{
\item{filename}{A data set saved in csv format with "," as field separator and "." as decimal
separator. Such file can be obtained from the .rep report provided by the instrument by opening
it with a spreadsheet editor and saving it as .csv file using a local setting = EN-US or
equivalent. The function identifies standard samples for bracketing when labelled as “SRM” and
substitutes their names with “SRM981_X” where X is a progressive integer. Similarly, QC samples
“CRM” are renamed as “CRM482_X”.}

\item{report}{A string referring to the type of report generated by the instrument Accepted
values are \code{"short"} or \code{"long"}. \itemize{ \item{\bold{Short}}{ reports have only
information about the sample ID, data and time of the measure, a summary with average
intensities and finally, the intensity values for each replicate.} \item{\bold{Long}}{ reports
contain also information of the "short" counterpart and also some information about the
measurement conditions. Additionally, the summary section is reported after the intensity
values for each replicates. These reports are tipically produced when working in isotope ratio
mode.} }}
}
\value{
A data.table containing raw intensities and raw Pb isotope ratios for each sample.
  Provided isotope ratios are 208Pb/207Pb, 206Pb/207Pb, 208Pb/206Pb and 207Pb/206Pb. Note that
  these ratios have still to be inspected for outliers, summarised and corrected for mass bias.
}
\description{
Extracts raw 208Pb/207Pb, 206Pb/207Pb, 208Pb/206Pb and 207Pb/206Pb data from the raw .csv file
produced by Perkin Elmer Elan ICP-MS.
}
\examples{
file.short <- system.file("extdata", "sednya_2015.csv", package = "pbratios")
sednya.2015 <- extract_data(file.short, report = "short")

file.long <- system.file("extdata", "spmnya_2012.csv", package = "pbratios")
spmnya.2012 <- extract_data(file.long, report = "long")

}
