% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoplot_generic.R
\name{isoplot.generic}
\alias{isoplot.generic}
\title{Generates a three-isotope plot.}
\usage{
isoplot.generic(data, x = "Pb208207", y = "Pb206207", factor = NULL,
  regression = FALSE)
}
\arguments{
\item{data}{A data.table generated by \code{corr.mbf} function and subsetted to
cointain only samples to be plotted in a three-isotope plot. Data should contain
the columns \code{Pb208207}, \code{Pb206207}, \code{Pb208206} and \code{Pb207206}
with the  measured Pb isotope ratio values for 208Pb/207Pb, 206Pb/207Pb,
208Pb/206Pb and 207Pb/206Pb, respectively.}

\item{x, y}{The string for \code{"Pb20x20y"} to be plotted on the x and y-axis.
Default values are \code{"Pb208207"} and \code{"Pb206207"}, respectively.}

\item{factor}{A factor contained within the data.table that should be used to colour
datapoints. Default value is \code{NULL}.}

\item{regression}{A logic expression for adding the regression line and it's \eqn{R^2}
value to the plot. Default is \code{FALSE}.}
}
\value{
The function returns a plot printed on screen. Additionally, three files
  named "data_gen" are saved as .pdf, .png and .svg in the "output" folder.
}
\description{
The function generates a generic three-isotope plot and saves it in three different
format (pdf, png and svg).
}
\examples{
data(pm10nya)
isoplot.generic(pm10nya)

}
\seealso{
\code{\link{corr.mbf}}
\code{\link{isoplot.pm10}}
}
