% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoplot_generic.R
\name{isoplot_generic}
\alias{isoplot_generic}
\title{Customisable three-isotope plots.}
\usage{
isoplot_generic(data, x = "Pb208207", y = "Pb206207", factor = NULL,
  regression = FALSE, save = FALSE)
}
\arguments{
\item{data}{A data.table generated by \code{corr.mbf} function and subsetted to cointain only
samples to be plotted in the three-isotope diagram. Data should contain the columns
\code{Pb208207}, \code{Pb206207}, \code{Pb208206} and \code{Pb207206} with the  measured Pb
isotope ratios for 208Pb/207Pb, 206Pb/207Pb, 208Pb/206Pb and 207Pb/206Pb, respectively.}

\item{x, y}{The string \code{"Pb20x20y"} to be plotted on the x and y-axis. Default values are
\code{"Pb208207"} and \code{"Pb206207"}, respectively.}

\item{factor}{A factor within the data.table that can be used to differentiate data points with
different colours. Default value is \code{NULL}.}

\item{regression}{Logical. If \code{TRUE}, a regression line and its \eqn{R^2} value are added to
the plot. Default is \code{FALSE}.}

\item{save}{Logical. If \code{TRUE}, the plot is saved in the "output" folder with the file name
"data_gen" in PDF, PNG and SVG formats.}
}
\value{
The function prints a three-isotope plot on screen. Additionally, if \code{save = TRUE},
  three files named "data_gen" are saved in the "output" folder with the file name "data_gen" in
  PDF, PNG and SVG formats.
}
\description{
The function plots Pb isotope ratio data in a three-isotope diagram, calculates a regression line
and saves the plot in PDF, PNG and SVG formats.
}
\examples{
data(pm10nya)
isoplot_generic(pm10nya, save = FALSE)

}
\seealso{
\code{\link{corr_mbf}} \code{\link{isoplot_pm10}}
}
