% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoplot_pm10.R
\name{isoplot_pm10}
\alias{isoplot_pm10}
\title{Three-isotope plot for PM10 data.}
\usage{
isoplot_pm10(data, factor = NULL, save = FALSE)
}
\arguments{
\item{data}{A data.table generated by corr.mbf function and subsetted to cointain only samples to
be plotted in the three-isotope diagram. Data should contain a column the column
"\code{Pb208206}, \code{Pb207206} with the measured Pb isotope ratios for 208Pb/206Pb and
207Pb/206Pb, respectively.}

\item{factor}{A factor within the data.table that can be used differentiate data points with
different point shapes. Default value is \code{NULL}.}

\item{save}{Logical. If \code{TRUE}, the plot is saved in the "output" folder with the file name
"data_gen" in PDF, PNG and SVG formats.}
}
\value{
The function prints a three-isotope plot on screen. Additionally, if \code{save = TRUE},
  three files named "data_gen" are saved in the "output" folder with the file name "data_gen" in
  PDF, PNG and SVG formats.
}
\description{
The function prints a three-isotope plot on screen and saves the plot as pdf, png and svg.
Measured data are compared with Pb isotope ratios measured in PM10 samples from the Northern
Hemisphere. Such data are obtained from the dataset \code{pm10nh} from the package
\code{pbratios}.
}
\examples{
isoplot_pm10(pm10nya, save = FALSE)

}
\seealso{
\code{\link{pm10nh}} \code{\link{isoplot_generic}}
}
