% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_flag.R
\name{out_flag}
\alias{out_flag}
\title{Identifying possible outliers.}
\usage{
out_flag(x)
}
\arguments{
\item{x}{A vector of numeric data.}
}
\value{
A logical vector. \code{TRUE} indicates that the corresponding value is outside the
  \eqn{median ± 2.5 mad} range.
}
\description{
The function identifies possible outliers based on the \eqn{median ± 2.5 mad} range criterion.
This function is used internally by the \code{calc.ratios} function.
}
\examples{
a <- rnorm(100)
b <- out_flag(a)

}
\seealso{
\code{\link{calc_ratios}}
}
