% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm10nya-raw.R
\name{pm10nya-raw}
\alias{pm10nya-raw}
\title{Raw csv report with Pb isotope ratio data.}
\format{The file has 26988 rows and 7 columns.}
\source{
Processed data were published in Bazzano, Cappelletti, Udisti and Grotti, Long-range
  transport of atmospheric lead reaching Ny-Ålesund: Inter-annual and seasonal variations of
  potential source areas. Atmospheric Environment, 2016, 139, 11-19.
}
\description{
This raw report was obtained saving as .csv the original .rep file produced by the mass
spectrometer Perkin Elmer ELAN ICP-MS.
}
\details{
Field separator are "," and decimal separator are ".". Standards with known Pb isotope ratios are
used to correct the instrumental mass bias by the standard bracketing technique and are labelled
as \code{SRM}. QC data are BCR CRM 482 and are labelled as \code{CRM}. The dataset comprises PM10
samples collected from 2010 to 2014 at Ny-Ålensund (Svalbards, Norwegian Arctic).
}
\author{
Andrea Bazzano and Marco Grotti.
}
