% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{qc_check}
\alias{qc_check}
\title{Checking quality control criteria.}
\usage{
qc_check(data)
}
\arguments{
\item{data}{A data set of Pb isotope ratios obtained by the function \code{corr.mbf}. The dataset
must have at least one sample labelled as \code{CRM}. The CRMs are be assumed to be BCR CRM 482
with reference values reported by Cloquet, C., Carignan, J., Libourel, G., 2006. Atmospheric
pollutant dispersion around an urban area using trace metal concentrations and Pb isotopic
compositions in epiphytic lichens. Atmos. Environ. 40, 574–587.}
}
\value{
The function returns a list containing \itemize{ \item{\code{data}}{ Pb isotope ratios
  measured for BCR CRM 482 samples.} \item{\code{summary}}{ Summary statistics from "data".}
  \item{\code{error}}{ Summary statistics for relative errors of experimental data compared to
  the reference values. Errors are reported on a permil scale.} Additionally, the function plots
  data points or boxplots (for \eqn{N > 10}) compared with reference values. }
}
\description{
The function estimates the performances of the analytical method by calculating basic descriptive
statistics on Pb isotope ratio data measured in BCR CRM 482 samples.
}
\examples{
data(pm10nya)
print(qc_check(pm10nya))

}
\seealso{
\code{\link{corr_mbf}}
}
