% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_table.R
\name{save_table}
\alias{save_table}
\title{Write a dataset into a .csv file.}
\usage{
save_table(data)
}
\arguments{
\item{data}{A dataset in the class data.frame or data.table.}
}
\value{
The function save a .csv file in the "output" folder. The filename is
  \code{data} without dots (".").
}
\description{
The function write a dataset into a .csv file with locale IT-it or equivalent.
}
\examples{
\dontrun{save_table(pm10nya)}  # save spmnya2012.csv in the "output" folder.

}
