% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outliers.R
\name{check_outliers}
\alias{check_outliers}
\title{Inspecting Pb isotope ratio replicates.}
\usage{
check_outliers(data, idname)
}
\arguments{
\item{data}{A data.table produced by the function \code{extract.data}. The dataset should contain
a \code{sample} column with sample names, \code{Pb20x} columns (\eqn{x = 6, 7, 8}) with the
measured intensities for each isotope, \code{Pb20x20y} columns storing 20xPb/20yPb raw isotope
ratios (\eqn{x, y = 6, 7, 8}).}

\item{idname}{A string identical to a sample name. (i.e. \code{"P40 I5"}).}
}
\value{
The function returns a list containing \itemize{ \item{\code{intensities}}{ Raw
  intensities for the monitored isotopes.} \item{\code{ratios}}{ Raw Pb isotope ratios with flag
  = \code{TRUE} if the value falls outside the \eqn{median ± 2.5 mad} range}. } Additionally, the
  function returns a plot with a panel for each Pb isotope ratio showing data points for each
  replicates. The green line is the median, yellow lines represent \eqn{median ± 2 mad} and red
  lines \eqn{median ± 3 mad}. Values inside the \eqn{median ± 2.5 mad} range are represented as
  points, possible outliers as *.
}
\description{
The function provides visual inspection of raw Pb isotope ratio replicates for each sample.
Possible outliers can be easily identified.
}
\examples{
file.long <- system.file("extdata", "spmnya_2012.csv", package = "pbratios")
spmnya.2012 <- extract_data(file.long, report = "long")
check_outliers(spmnya.2012, "II.B.50m")

}
\seealso{
\code{\link{calc_ratios}}
}
