% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoplot_pm10.R
\name{isoplot_pm10}
\alias{isoplot_pm10}
\title{Three-isotope plot for PM10 data.}
\usage{
isoplot_pm10(data, factor = NULL)
}
\arguments{
\item{data}{A data.table generated by corr.mbf function and subsetted to cointain only
samples to be plotted in a three-isotope plot. Data should contain a column
the column "\code{Pb208206}, \code{Pb207206} with the measured Pb isotope ratio
 values for 208Pb/206Pb and 207Pb/206Pb, respectively.}

\item{factor}{A factor contained within the data.table and used to differentiate
datapoints using different point geometries. Default value is \code{NULL}.}
}
\value{
The function prints a plot on screen and saves three files as .pdf, .png
  and .svg in the "output" folder.
}
\description{
The function prints a three-isotope plot on screen and saves the plot as pdf, png and
svg. Measured data are compared with Pb isotope ratios measured in PM10 samples from
the Northern Hemisphere. Such data are obtained from the dataset \code{pm10nh} from
the package \code{pbratios}.
}
\examples{
isoplot_pm10(pm10nya)

}
\seealso{
\code{\link{pm10nh}}
\code{\link{isoplot_generic}}
}
