% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbratios.R
\docType{package}
\name{pbratios}
\alias{pbratios}
\alias{pbratios-package}
\title{Processing Pb isotope ratio data.}
\description{
The package provides some functions to efficiently extract and process Pb isotope ratio
data acquired with a Perkin Elmer ELAN ICP-MS working in isotope ratio mode.
}
\details{
The principal functions are
\describe{
\item{\code{extract_data}}{ Extracts data from the csv report produced by the mass
  spectrometer. 208Pb/207Pb, 206Pb/207Pb, 208Pb/206Pb and 207Pb/206Pb ratios
  will be taken into account.}
\item{\code{calc_ratios}}{ Remove outliers from the dataset and summarise the
  replicates.}
\item{\code{corr_mbf}}{ Corrects the Pb isotope ratio values for instrumental mass
  fractionation usingh the bracketing standard technique.}
}
Other functions are then provided for inspecting possible outliers, evaluating QC
samples, saving the final dataset and produce some three-isotope plots.
}
\seealso{
\code{\link{extract_data}}
  \code{\link{calc_ratios}}
  \code{\link{corr_mbf}}
}
\author{
Andrea Bazzano <andrea.bazzano86@gmail.com>
}
