% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{qc_check}
\alias{qc_check}
\title{Checking quality control criteria.}
\usage{
qc_check(data)
}
\arguments{
\item{data}{a dataset of Pb isotope ratio values obtained by the function
\code{corr.mbf}.
The dataset must contain at least one sample labelled as \code{CRM}.
The CRM will be assumed as BCR CRM 482 and values reported by
Cloquet, C., Carignan, J., Libourel, G., 2006. Atmospheric pollutant
dispersion around an urban area using trace metal concentrations and Pb
isotopic compositions in epiphytic lichens. Atmos. Environ. 40, 574–587
will be considered as reference values.}
}
\value{
The function returns a list containing
   \itemize{
   \item{\code{data}}{ Corrected Pb isotope ratios measured for BCR CRM 482 QC
     samples.}
   \item{\code{summary}}{ Summary statistics from "data".}
   \item{\code{error}}{ Summary statistics for experimental data vs reference values.
      Errors are reported on a permil scale.}
   Additionally, the function plots the datapoints or boxplots (for \eqn{N > 10}).
}
}
\description{
The function checks quality control criteria for Pb isotope ratios measured in BCR CRM
482 QC samples.
}
\examples{
data(pm10nya)
print(qc_check(pm10nya))

}
\seealso{
\code{\link{corr_mbf}}
}
